tracing_journald

Enum Priority

#[repr(u8)]
pub enum Priority {
    Emergency = 48,
    Alert = 49,
    Critical = 50,
    Error = 51,
    Warning = 52,
    Notice = 53,
    Informational = 54,
    Debug = 55,
}

A priority (called โseverity codeโ by syslog) is used to mark the importance of a message.

Descriptions and examples are taken from the Arch Linux wiki. Priorities are also documented in the section 6.2.1 of the Syslog protocol RFC.

Variants

Emergency = 48

System is unusable.

Examples:

severe Kernel BUG

systemd dumped core

This level should not be used by applications.

Alert = 49

Should be corrected immediately.

Examples:

Vital subsystem goes out of work, data loss:

kernel: BUG: unable to handle kernel paging request at ffffc90403238ffc

Critical = 50

Critical conditions

Examples:

Crashe, coredumps

systemd-coredump[25319]: Process 25310 (plugin-container) of user 1000 dumped core

Error = 51

Error conditions

Examples:

Not severe error reported

kernel: usb 1-3: 3:1: cannot get freq at ep 0x84, systemd[1]: Failed unmounting /var

libvirtd[1720]: internal error: Failed to initialize a valid firewall backend

Warning = 52

May indicate that an error will occur if action is not taken.

Examples:

a non-root file system has only 1GB free

org.freedesktop. Notifications[1860]: (process:5999): Gtk-WARNING **: Locale not supported by C library. Using the fallback 'C' locale

Notice = 53

Events that are unusual, but not error conditions.

Examples:

systemd[1]: var.mount: Directory /var to mount over is not empty, mounting anyway

gcr-prompter[4997]: Gtk: GtkDialog mapped without a transient parent. This is discouraged

Informational = 54

Normal operational messages that require no action.

Example: lvm[585]: 7 logical volume(s) in volume group "archvg" now active

Debug = 55

Information useful to developers for debugging the application.

Example: kdeinit5[1900]: powerdevil: Scheduling inhibition from ":1.14" "firefox" with cookie 13 and reason "screen"