Crate tracing_journald

tracing-journald

Support for logging tracing events natively to journald, preserving structured information.

Overview

tracing is a framework for instrumenting Rust programs to collect scoped, structured, and async-aware diagnostics. tracing-journald provides a [tracing-subscriber::Layer] implementation for logging tracing spans and events to systemd-journald, on Linux distributions that use systemd.

Compiler support: requires rustc 1.49+

Supported Rust Versions

Tracing is built against the latest stable release. The minimum supported version is 1.49. The current Tracing version is not guaranteed to build on Rust versions earlier than the minimum supported version.

Tracing follows the same compiler support policies as the rest of the Tokio project. The current stable Rust compiler and the three most recent minor versions before it will always be supported. For example, if the current stable compiler version is 1.45, the minimum supported version will not be increased past 1.42, three minor versions prior. Increasing the minimum supported compiler version is not considered a semver breaking change as long as doing so complies with this policy.

Structs

Layer

Sends events and their fields to journald

Functions

layer

Construct a journald layer